refactor(mattermost): share wildcard callback host check
